my-multi-vendor-ecommerce/backend/requirements.txt
Django==4.1.0
djangorestframework==3.13.1
requests==2.27.1
hubspot-api-client==3.0.0
python-dotenv==0.19.2
stripe==2.62.0
django-cors-headers==3.13.0
django-environ==0.8.1